Add Security

+++ src/main/java/com/library/service/impl/LibrarianServiceImpl.java
@@ -1,18 +1,11 @@
 package com.library.service.impl;
 
 import com.library.controller.exception.BookNotFoundException;
+import com.library.controller.exception.BookOutOfStockException;
 import com.library.controller.exception.PersonNotFoundException;
-import com.library.domain.dto.BookInfoDto;
-import com.library.domain.dto.PersonDto;
-import com.library.domain.dto.TakenBookDto;
-import com.library.domain.entity.Author;
-import com.library.domain.entity.Book;
-import com.library.domain.entity.Person;
-import com.library.domain.entity.TakenBook;
-import com.library.repository.AuthorRepository;
-import com.library.repository.BookRepository;
-import com.library.repository.PersonRepository;
-import com.library.repository.TakenBookRepository;
+import com.library.domain.dto.*;
+import com.library.domain.entity.*;
+import com.library.repository.*;
 import com.library.service.LibrarianService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.convert.ConversionService;
@@ -31,6 +24,7 @@ public class LibrarianServiceImpl implements LibrarianService {
     private final BookRepository bookRepository;
     private final AuthorRepository authorRepository;
     private final TakenBookRepository takenBookRepository;
+    private final WarehouseRepository warehouseRepository;
     private final ConversionService conversionService;
 
     @Override
@@ -41,7 +35,7 @@ public class LibrarianServiceImpl implements LibrarianService {
     }
 
     @Override
-    public BookInfoDto createBook(BookInfoDto bookInfoDto) {
+    public BookInfoWithPlaceDto createBook(BookInfoDto bookInfoDto) {
         Book book = conversionService.convert(bookInfoDto, Book.class);
         Set<Author> authors = book.getAuthors();
         for(Author author : authors){
@@ -52,25 +46,37 @@ public class LibrarianServiceImpl implements LibrarianService {
         }
         book.setAuthors(authors);
         Book result = bookRepository.save(book);
-        return conversionService.convert(result, BookInfoDto.class);
+        Warehouse warehouse = new Warehouse(result, (int)(1+Math.random()*20), (int)(1+Math.random()*20), 1);
+        Warehouse warehouseResult = warehouseRepository.save(warehouse);
+        return conversionService.convert(result, BookInfoWithPlaceDto.class);
     }
 
     @Override
-    public TakenBookDto addBook(UUID personId, UUID bookId){
+    public TakenBookWithPlaceDto addBook(UUID personId, UUID bookId){
         Optional<Person> person = personRepository.findById(personId);
         if (person.isEmpty()) throw new PersonNotFoundException();
         Optional<Book> book = bookRepository.findById(bookId);
         if (book.isEmpty()) throw new BookNotFoundException();
+        if(warehouseRepository.getNumber(bookId) < 1) throw new BookOutOfStockException();
         TakenBook takenBook = new TakenBook();
         takenBook.setPerson(person.get());
         takenBook.setBook(book.get());
         takenBook.setDateOfReceiving(LocalDate.now());
         TakenBook result = takenBookRepository.save(takenBook);
-        return conversionService.convert(result, TakenBookDto.class);
+        warehouseRepository.decreaseNumber(bookId);
+        return conversionService.convert(result, TakenBookWithPlaceDto.class);
     }
 
     @Override
     public void removeBook(UUID id){
+        Optional<TakenBook> takenBook = takenBookRepository.findById(id);
+        if (takenBook.isEmpty()) throw new BookNotFoundException();
         takenBookRepository.deleteById(id);
+        warehouseRepository.increaseNumber(takenBook.get().getBook().getId());
+    }
+
+    @Override
+    public Warehouse findWarehouseByBookId(UUID bookId) {
+        return warehouseRepository.findByBookId(bookId);
     }
 }
